requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
backports.functools-lru-cache==1.5  # via cheroot
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
cheroot==6.5.5            # via cherrypy
cherrypy==18.1.2
configobj==5.0.6
google-api-core[grpc]==1.14.0  # via google-cloud-speech
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.6.3        # via google-api-core, google-api-python-client, google-auth-httplib2
google-cloud-speech==1.1.0
googleapis-common-protos==1.6.0  # via google-api-core
grpcio==1.22.0            # via google-api-core
httplib2==0.13.0
idna==2.8                 # via requests
jaraco.functools==2.0     # via tempora
mako==1.0.14
markupsafe==1.1.1         # via mako
more-itertools==7.2.0     # via cheroot, cherrypy, jaraco.functools
passlib==1.7.1
portend==2.5
protobuf==3.9.0           # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.5     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
pyaudio==0.2.11
pyjwt==1.7.1
python-dateutil==2.8.0
pytz==2019.1              # via google-api-core, tempora, tzlocal
pywin32==224 ; sys_platform == "win32"
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
six==1.12.0               # via cheroot, configobj, google-api-core, google-api-python-client, google-auth, grpcio, protobuf, python-dateutil, tempora
tempora==1.14.1           # via portend
tzlocal==2.0.0
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.25.3           # via requests
zc.lockfile==1.4          # via cherrypy

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via google-api-core, protobuf, zc.lockfile
